requirements.txt
altgraph==0.17.3
anyio==3.6.2
certifi==2022.12.7
click==8.1.3
Flask==2.2.3
greenlet==2.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.1
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
macholib==1.16.2
MarkupSafe==2.1.2
Pillow==9.4.0
playwright==1.29.0
pyee==9.0.4
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
python-dotenv==0.21.0
rfc3986==1.5.0
sniffio==1.3.0
typing_extensions==4.4.0
Werkzeug==2.2.3
zipp==3.15.0
